:sparkles: feat: 페이지 정보 가져오기 구현

+++ backend/src/main/java/com/donjo/backend/db/entity/Support.java
@@ -31,7 +31,7 @@ public class Support {
     private String supportType;
 
     @Column
-    private Integer supportUid;
+    private Long supportUid;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "from_address", referencedColumnName = "address")
@@ -40,7 +40,6 @@ public class Support {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "to_address", referencedColumnName = "address")
     private Member toAddress;
-
 
     @Column
     private String sendMsg;
